# New heuristic based on lower/upper bounds, key=None argument