Thanks @eivindjahren

**Before**

```
Benchmarking: anytime_beam_search(board, {'power': 10})
 ..........
 Solutions of avg. length 15.000 in 17.216 sec.
 Avg. length times sec (lower is better) => 258.234

Benchmarking: heuristic_search(board, {'max_nodes': 150000})
 ..........
 Solutions of avg. length 17.300 in 17.647 sec.
 Avg. length times sec (lower is better) => 305.293

Benchmarking: monte_carlo_search(board, {'iterations': 3320, 'seed': 1})
 ..........
 Solutions of avg. length 16.200 in 15.334 sec.
 Avg. length times sec (lower is better) => 248.413
```

**After**

```
Benchmarking: best_first_search(board, {'exponent': None, 'seed': 0})
 ..........
 Solutions of avg. length 17.400 in 0.021 sec.
 Avg. length times sec (lower is better) => 0.360

Benchmarking: anytime_beam_search(board, {'power': 9})
 ..........
 Solutions of avg. length 14.000 in 13.699 sec.
 Avg. length times sec (lower is better) => 191.792

Benchmarking: heuristic_search(board, {'max_nodes': 64000, 'verbose': False})
 ..........
 Solutions of avg. length 15.700 in 14.024 sec.
 Avg. length times sec (lower is better) => 220.179

Benchmarking: monte_carlo_search(board, {'iterations': 650, 'seed': 1})
 ..........
 Solutions of avg. length 15.200 in 14.655 sec.
 Avg. length times sec (lower is better) => 222.752
```